+++ convex/staticHosting.ts
@@ -9,8 +9,14 @@ import {
 } from "@convex-dev/static-hosting";
 import { components } from "./_generated/api";
 
-export const { generateUploadUrl, recordAsset, gcOldAssets, listAssets } =
-  exposeUploadApi(components.selfHosting);
+export const {
+  generateUploadUrl,
+  generateUploadUrls,
+  recordAsset,
+  recordAssets,
+  gcOldAssets,
+  listAssets,
+} = exposeUploadApi(components.selfHosting);
 
 export const { getCurrentDeployment } = exposeDeploymentQuery(
   components.selfHosting,